requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.7=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py310h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.2.0=py310hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.1=h6a678d5_0
idna=3.4=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py310h06a4308_0
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.8.0=py310h06a4308_0
matplotlib-base=3.8.0=py310h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
numexpr=2.8.7=py310h85018f9_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=py310h06a4308_0
pandas=2.2.1=py310h6a678d5_0
pcre2=10.42=hebb0a14_0
pillow=10.2.0=py310h5eee18b_0
pip=23.3.1=py310h06a4308_0
ply=3.11=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.10=py310h6a678d5_0
pyqt5-sip=12.13.0=py310h5eee18b_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.1=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py310h06a4308_0
pyyaml=6.0.1=py310h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
requests=2.31.0=py310h06a4308_1
scikit-learn=1.3.0=py310h1128e8f_1
scipy=1.12.0=py310h5f9d8c6_0
seaborn=0.12.2=py310h06a4308_0
setuptools=68.2.2=py310h06a4308_0
sip=6.7.12=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py310h06a4308_0
torchaudio=2.2.1=py310_cu121
torchtriton=2.2.0=py310
torchvision=0.17.1=py310_cu121
tornado=6.3.3=py310h5eee18b_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py310h06a4308_1
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
